Update XMLDeserializer.php
You should not cast to a number here. I have XML that has a productnumber <productid>00012332</productid>. Casting to an int means  the leading zeros get chopped. I also think is_numeric will be true, if the node's content is some decimal (123.345) and you cast that to an int - that will give you 123. Maybe the same problem applies to json-deserializer?